This commit will,
Change the main script (paraevaluar), so that executes different bash
scripts for each given functionality. The functionalities are
initialization, job script creation, evaluation, and reporting.

Add the bash script for initialization that creates the directory and
all the needed files.

Add the bash script for the job script creation.

Add the bash script for the evaluation, which does the actual execution
of the evaluated binaries.

Add the template for the configuration file of the script which is used
in the evaluation and report stages.